% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_tos_database}
\alias{get_tos_database}
\title{get_tos_database.R
extract .csv for plant, vertebrate, or invertebrate ToS dbs and rename columns<\n>}
\usage{
get_tos_database(tos_db)
}
\arguments{
\item{tos_db}{ToS database to download, either "plant", "vertebrate", or "invertebrate"}
}
\value{
ToS database with reformatted headers
}
\description{
get_tos_database.R
extract .csv for plant, vertebrate, or invertebrate ToS dbs and rename columns<\n>
}
\examples{
plant_tos_db <- get_tos_database("plant")
}

