% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_tosr_load}
\alias{my_tosr_load}
\title{A list with data to create ToS}
\format{
A list with two dataset
}
\source{
\url{https://www.scopus.com} \url{https://www.webofknowledge.com}
}
\usage{
my_tosr_load
}
\description{
A list of three dataset that contain:
\enumerate{
\item Graph: this is a citation network build with the references
\item df: A tibble with the main information of our initial
dataset with some more ids
\item nodes: A tibble with the complete information of each node
from graph dataset (1).
}
}
\keyword{datasets}
