% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_trait_dendrogram}
\alias{plot_trait_dendrogram}
\title{Plot a phylogeny with discrete character traits
Description: Plots a phylogeny with binary or discrete character states, given a tree and a trait matrix}
\usage{
plot_trait_dendrogram(apetree, character_vector, tree_type = "phylogram",
  binary_trait = FALSE, legend = TRUE, legend_cex = 0.75,
  tiplabel_cex = 0.4, legend_loc = "bottomleft")
}
\arguments{
\item{apetree}{(an ape tree object)}

\item{tree_type}{(what class of tree to build; defaults to phylogram)}

\item{character_matrix}{(a data frame of species:trait relationship)}

\item{header}{(the dataframe column header of your character of interest)}
}
\value{
diversitree plot with color-coded characters as tips
}
\description{
Plot a phylogeny with discrete character traits
Description: Plots a phylogeny with binary or discrete character states, given a tree and a trait matrix
}
\examples{
plot_trait_dendrogram(tree, charactermatrix, tree_type="fan")
}
\author{
Alex Harkess
}

