% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr.R
\name{tosr}
\alias{tosr}
\title{tosr process}
\usage{
tosr(files, layout)
}
\arguments{
\item{files}{a data frame with ToS data}
}
\value{
list with graph, data frame of citations, subfields and Tree of science
}
\description{
Load of files, building of graph, and SAP algorithm for Tree of science
}
\author{
Sebastian Robledo
}
