% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_bibliometix_analysis.R
\name{tosr_bibliometix_analysis}
\alias{tosr_bibliometix_analysis}
\title{bibliometric analysis}
\usage{
tosr(df)
}
\arguments{
\item{bibliometrix_objetc}{a bibliometix objetc}
}
\value{
bibliometrix analysis plots
}
\description{
Compute the bibliometric analysis using bibliometrix library
}
\author{
Sebastian Robledo
}
