% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_graph.R
\name{tosr_graph}
\alias{tosr_graph}
\title{Creating a graph object}
\usage{
tosgraph(data = scopus_dataframe)
}
\arguments{
\item{scopus_dataframe}{a dataframe with scopus data}
}
\value{
graph: a graph object
}
\description{
Load a dataframe with scopus .bib data and convert it in a graph
}
\examples{
library(treeofscience)
          socopus_df   <- tosload("scopus.bib") # Create scopus data frame
          graph.scopus <- tosgraph(scopus_df) # Create graph from scopus data frame
}
\author{
Sebastian Robledo
}
