% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_load.R
\name{tosr_load}
\alias{tosr_load}
\title{Loading .bib files}
\usage{
tosload(file = name of file)
}
\arguments{
\item{file.bib}{the .bib file to be converted}
}
\value{
dataframe with all variables in columns
}
\description{
load .bib file and convert it to data frame with scupos information
}
\examples{
library(treeofscience)
          file         <- "scopus.bib"
          socopus_df   <- tosload(file) # Create scopus data frame
}
\author{
Sebastian Robledo
}
