% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_process.R
\name{tosr_process}
\alias{tosr_process}
\title{Creating our ToS}
\usage{
tosprocess(graph)
}
\arguments{
\item{graph}{a graph object with scopus data}
}
\value{
a dataframe with the Tree of science
}
\description{
Convert a graph object into the tree of science
}
\examples{
library(treeofscience)
          socopus_df <- tosload("scopus.bib") # Create scopus data frame
          graph.scopus <- tosgraph(scopus_df) # Create graph from scopus data frame
          TOS <- tosprocess(graph.scopus)     # Create Tree of Science
}
\author{
Sebastian Robledo
}
