% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_visualization.R
\name{tosr_visualization}
\alias{tosr_visualization}
\alias{tosvisualization}
\title{Visualization of TOS information}
\usage{
tosvisualization(graph.tos, TOS, layout)
}
\arguments{
\item{tosdata, layout}{a data frame with ToS data}
}
\value{
Visualization
}
\description{
Create a visualization of the Tree of Science using visNetwork
}
\author{
Sebastian Robledo
}
