% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosSAP.R
\name{tosSAP}
\alias{tosSAP}
\title{SAP algorithm for tree of science}
\usage{
tosSAP(graph,df,nodes)
}
\value{
Dataframe with tree of science
}
\description{
Compute tree of science
}
\author{
Sebastian Robledo
}
