% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosR.R
\name{tosR}
\alias{tosR}
\title{tosR process}
\usage{
tosR(...)
}
\arguments{
\item{...}{Names of scopus and web of science files}
}
\value{
list with graph, data frame of citations, subfields and Tree of science
}
\description{
This function load de scopus and web of science files and generate
the graph (relations betwen the cited references), the bibliometrix
dataframe object, the tree of science for the subfields, the tree of
science for all the cited references,
}
\examples{
\dontrun{
my_tosr <- tosR("co-citation_209.txt",
               "co-citation_380.bib")
}
}
\author{
Sebastian Robledo
}
