% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_bibliometix_analysis.R
\name{tosr_bibliometix_analysis}
\alias{tosr_bibliometix_analysis}
\title{Bibliometric analysis}
\usage{
tosr_bibliometix_analysis(df)
}
\arguments{
\item{df}{A bibliometix objetc}
}
\value{
bibliometrix analysis plots
}
\description{
Bibliometric analysis using bibliometrix library
}
\examples{
library(tosr)
tosinfo <- tosR('scopus.bib','save.txt') # Tos process
Summary <- tosr_bibliometix_analysis(tosinfo.df)

}
\author{
Sebastian Robledo
}
