% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosr_process.R
\name{tosr_process}
\alias{tosr_process}
\title{Tree of science for subfields}
\usage{
TOS_subfields <- tosr_process(g, df, nodes, number_nodes)
}
\arguments{
\item{graph}{Graph object with bibliometrix data obtained from function tosr_load}

\item{df}{Bibliometrix data frame obtained from function tosr_load}

\item{nodes}{Dataframe with nodes atributes obtained from function tosr_load}

\item{number_nodes}{Scalar number for minimum number of nodes to be considered in the subfield}
}
\value{
Dataframe with the Tree of science for subfields
}
\description{
Convert a graph object into the tree of science
}
\examples{
library(tosr)
data_info     <- tosload("scopus.bib") # Create scopus data frame
TOS_subfields <- tosr_process(g  = data_info$g,
                              df = data_info$df,
                           nodes = data_info$nodes,
                    number_nodes = 50)
}
\author{
Sebastian Robledo
}
