% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosSAP.R
\name{tosSAP}
\alias{tosSAP}
\title{SAP algorithm for tree of science}
\usage{
TOS <- tosSAP(graph,df,nodes)
}
\arguments{
\item{graph}{A graph object with bibliometrix data obtained from function tosr_load}

\item{df}{Bibliometrix data frame obtained from function tosr_load}

\item{nodes}{Dataframe with nodes atributes obtained from function tosr_load}
}
\value{
Dataframe with tree of science
}
\description{
Compute tree of science using SAP algorithm.
             see: https://revistas.unal.edu.co/index.php/ingeinv/article/view/77718/0
}
\author{
Sebastian Robledo
}
