% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosSAP.R
\name{tosSAP}
\alias{tosSAP}
\title{SAP algorithm for tree of science}
\usage{
tosSAP(graph, df, nodes)
}
\arguments{
\item{graph}{A graph object with bibliometrix data obtained from function tosr_load}

\item{df}{Bibliometrix data frame obtained from function tosr_load}

\item{nodes}{Dataframe with nodes atributes obtained from function tosr_load}
}
\value{
Dataframe with tree of science
}
\description{
Compute tree of science using SAP algorithm.
see: https://revistas.unal.edu.co/index.php/ingeinv/article/view/77718/0
}
\examples{
\dontrun{
my_tos <- tosSAP(my_tos_load$graph, my_tos_load$df, my_tos_load$nodes)
}
}
\author{
Sebastian Robledo
}
